Create utils to refactor index
