+++ packages/datadog-plugin-kafkajs/src/index.js
@@ -1,6 +1,14 @@
 'use strict'
+const {
+  producer: {
+    addCommonProducerTags,
+    createProducerRequestQueueSizeTags,
+    createProducerRequestTags,
+    createProducerRequestTimeoutTags
+  }
+} = require('./utils')
 
-function createWrapRequest (tracer, config) {
+function createWrapProducer (tracer, config) {
   return function wrapProducer (createProducer) {
     return function producerWithTrace () {
       const serviceName = config.service || `${tracer._service}-kafka`
@@ -59,83 +67,22 @@ function createWrapRequest (tracer, config) {
   }
 }
 
-function createProducerRequestTimeoutTags ({
-  broker,
-  clientId,
-  correlationId,
-  createdAt,
-  sentAt,
-  pendingDuration,
-  apiName,
-  apiVersion
-}) {
-  return {
-    'kafka.broker': broker,
-    'kafka.clientId': clientId,
-    'kafka.correlationId': correlationId,
-    'kafka.message.createdAt': createdAt,
-    'kafka.message.sentAt': sentAt,
-    'kafka.pendingDuration': pendingDuration,
-    'kafka.apiName': apiName,
-    'kafka.apiVersion': apiVersion
-  }
-}
-
-function createProducerRequestQueueSizeTags ({ broker, clientId, queueSize }) {
-  return {
-    'kafka.broker': broker,
-    'kafka.clientId': clientId,
-    'kafka.queueSize': queueSize
-  }
-}
-// eslint-disable-next-line max-len
-function createProducerRequestTags ({
-  broker,
-  clientId,
-  correlationId,
-  size,
-  createdAt,
-  sentAt,
-  pendingDuration,
-  duration,
-  apiName,
-  apiVersion
-}) {
-  return {
-    'kafka.broker': broker,
-    'kafka.clientId': clientId,
-    'kafka.correlationId': correlationId,
-    'kafka.message.size': size,
-    'kafka.message.createdAt': createdAt,
-    'kafka.message.sentAt': sentAt,
-    'kafka.pendingDuration': pendingDuration,
-    'kafka.duration': duration,
-    'kafka.apiName': apiName,
-    'kafka.apiVersion': apiVersion
-  }
-}
-
-function addCommonProducerTags (serviceName, resourceName, tagCreatorFn) {
-  const restOfTags = tagCreatorFn ? tagCreatorFn() : {}
-
-  return {
-    'service.name': serviceName,
-    'resource.name': resourceName,
-    'span.kind': 'producer',
-    'span.type': 'queue',
-    component: 'kafkajs',
-    ...restOfTags
-  }
-}
-
 module.exports = [
   {
     name: 'kafkajs',
     versions: ['>=1.2'],
     patch ({ Kafka }, tracer, config) {
-      this.wrap(Kafka.prototype, 'producer', createWrapRequest(tracer, config))
-      this.wrap(Kafka.prototype, 'consumer', createWrapRequest(tracer, config))
-      this.wrap(Kafka.prototype, 'admin', createWrapRequest(tracer, config))
+      this.wrap(
+        Kafka.prototype,
+        'producer',
+        createWrapProducer(tracer, config)
+      )
+      this.wrap(
+        Kafka.prototype,
+        'consumer',
+        createWrapProducer(tracer, config)
+      )
+      this.wrap(Kafka.prototype, 'admin', createWrapProducer(tracer, config))
     },
     unpatch ({ Kafka }) {
       this.unwrap(Kafka.prototype, 'producer')
